+++ internal/infrastructure/database/postgres/repositories/shipment_test.go
@@ -12,6 +12,7 @@ import (
 	"github.com/emoss08/trenova/internal/core/domain/shipmenttype"
 	"github.com/emoss08/trenova/internal/core/domain/user"
 	"github.com/emoss08/trenova/internal/core/ports"
+	"github.com/emoss08/trenova/internal/core/services/calculator"
 	"github.com/stretchr/testify/require"
 
 	repoports "github.com/emoss08/trenova/internal/core/ports/repositories"
@@ -41,6 +42,7 @@ func TestShipmentRepository(t *testing.T) {
 		Logger:        logger.NewLogger(testutils.NewTestConfig()),
 		DB:            ts.DB,
 		ProNumberRepo: proNumberRepo,
+		Calculator:    calculator.NewShipmentCalculator(calculator.ShipmentCalculatorParams{Logger: logger.NewLogger(testutils.NewTestConfig())}),
 	})
 
 	t.Run("list shipments", func(t *testing.T) {
