+++ ui/src/app/shipment/_components/form/move-assignment-details.tsx
