+++ ui/src/app/shipment/_components/form/shipment-detail-column.tsx
