+++ ui/src/app/shipment/_components/form/shipment-details-components.tsx
